Separate out clientsfilter component
