+++ client/src/components/Clients/sort.test.js
@@ -1,4 +1,4 @@
-import { sortByProperty } from ".";
+import { sortByProperty } from "./index";
 
 const testClients = [
   {
